feat(models): refresh Gemini and Cohere catalogs
