+++ src/core/providers/ollama/streaming.rs
@@ -316,6 +316,7 @@ fn response_to_chunks(response: ChatResponse) -> Vec<ChatChunk> {
             delta: ChatDelta {
                 role: Some(MessageRole::Assistant),
                 content: None,
+                audio: None,
                 thinking: None,
                 tool_calls: None,
                 function_call: None,
@@ -352,6 +353,7 @@ fn response_to_chunks(response: ChatResponse) -> Vec<ChatChunk> {
                         delta: ChatDelta {
                             role: None,
                             content: Some(chunk_text),
+                            audio: None,
                             thinking: None,
                             tool_calls: None,
                             function_call: None,
@@ -376,6 +378,7 @@ fn response_to_chunks(response: ChatResponse) -> Vec<ChatChunk> {
                 delta: ChatDelta {
                     role: None,
                     content: None,
+                    audio: None,
                     thinking: None,
                     tool_calls: None,
                     function_call: None,
